requirements.txt
matplotlib==1.4.3
Pillow==5.1.0
